+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Gracefully handle a missing table name when reading an XER file.
 * Gracefully handle a unexpected calendar data when reading an XER file.
 * Correctly handle XER files with multi-byte character encoding.
+* Ensure project calendars are read from PMXML files.
 * Added readAll methods to PrimaveraPMFileWriter to allow all projects contained in a PMXML file to be read in a single pass.
 
 ## 8.1.3 (25/06/2020)